+++ example/index.tsx
@@ -46,6 +46,8 @@ const mountManager = async (room, root) => {
         // fullscreen: true,
         debug: true,
         cursor,
+        cursorOptions: { style: "custom" },
+        overwriteStyles: ".netless-window-manager-cursor-name { display: none }",
     })) as WindowManagerType;
 
     manager.emitter.on("ready", async () => {
@@ -235,7 +237,7 @@ const App = () => {
                     height: "calc(100vh - 32px)",
                     border: "1px solid",
                     resize: "auto",
-                    overflow: "scroll"
+                    overflow: "scroll",
                 }}
             ></div>
             <div className="side">